requirements.txt
"""
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pillow>=8.3.0
tqdm>=4.62.0
tensorboard>=2.7.0
scikit-image>=0.18.0
matplotlib>=3.4.0
opencv-python>=4.5.0
nibabel>=3.2.0
SimpleITK>=2.1.0
"""